Cloud_Project_1/requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
django-filter==2.2.0
pytz==2019.2
sqlparse==0.3.0
